fix: getStatusTextをswitch文にして共通化
